+++ src/Fable.EdIlyin.Core/Decode.fs
@@ -211,3 +211,6 @@ let resultFromResult =
 
 let fromDecodeResult decodeResult =
     primitive "result" <| fun _ -> decodeResult
+
+
+let